Fix crash in UE 5.1.1 when importing new texture (#2)
This fixes issue https://github.com/benui-dev/UE-BUIValidator/issues/4
by checking for pre-existing settings instead of trying to get a valid
Texture Factory. Because it would likely require bigger changes if it's
to work with the Interchange plugin. We check if the texture already has
some settings that aren't valid and if so set the default (first)
setting for each.