Add validation for `Filter` setting (#4)
The `Filter` setting is very important for pixel art and blocky UI
textures. It has to be set to `Nearest` to avoid blurriness when
scaling. Other games might want other settings, so the default value is
`TF_Default` (`"Default (from Texture Group)"`):


![image](https://github.com/seriema/UE-BUIValidator/assets/693684/f5b5c349-c639-4dff-b929-8f8779af0e00)

It's under the Validation Rules:


![image](https://github.com/seriema/UE-BUIValidator/assets/693684/c746bd36-3915-4b9a-9b33-6bfaaa9e7e86)

> Note about `Nearest`: The art itself should follow the "Base-4 rule"
but validating that would require a very different tool than
BUIValidator. So assuming the UI artist made the art, then this
validator can check that the scaling will apply Nearest.